Remove legacy ligandModal file
