+++ src/modal/LigandModal.js
@@ -1,0 +1,28 @@
+import LigandDetails from './LigandDetails.js';
+import SimilarLigandTable from './SimilarLigandTable.js';
+import PdbEntryList from './PdbEntryList.js';
+
+class LigandModal {
+    constructor(moleculeManager) {
+        this.details = new LigandDetails(moleculeManager);
+        this.similarLigandTable = new SimilarLigandTable(moleculeManager);
+        this.pdbEntryList = new PdbEntryList(moleculeManager);
+    }
+
+    show(ccdCode, sdfData) {
+        this.details.show(ccdCode, sdfData);
+        this.similarLigandTable.load(ccdCode);
+        this.pdbEntryList.load(ccdCode);
+    }
+
+    close() {
+        this.details.close();
+    }
+
+    load2DStructure(ccdCode, container) {
+        return this.similarLigandTable.load2DStructure(ccdCode, container);
+    }
+}
+
+export default LigandModal;
+
